refactor: move some stateless logic of client to its own crate
Currently, use of the SDK necessitates use of a client, which performs
networking and IO.
However, some of the logic of the SDK can be defined statelessly.
Doing so has many benefits, but the most urgent one is being able to use
that logic in restricted environments.
Subsequent work will expose some of this logic with wasm bindings,
making it usable by javascript code.