feat: add crate with wasm bindings for stateless sdk logic
Currently, reading SxT Chain data is difficult outside of rust since
proof-of-sql is native to rust.
In particular, we'd like to be able to use the chain and perform
proof-of-sql verified queries in javascript for chainlink integration.
This crate provides bindings for the stateless functionality of the SDK,
so that js can form requests for and deal with responses from the chain.