feat: add javascript client and script to attach it to wasm-pack
A javascript port of the rust sdk client is convenient for use in
javascript environments like chainlink. This adds such a port in
index_tail.js, which can be attached to the index.js produced by
wasm-pack using the provided build_index.sh script.

This code was originally written by @JayWhite2357 and @Dustin-Ray.