fix: flatten node package
We have tested publishing npm packages for use in chainlink with flat
npm packages successfully. The recent release automation didn't have a
flat package, relying on a src directory instead. This package is not
able to be imported. While it's probably possible to make that structure
work, for now we will revert to something we have gotten to work in the
past.