fix: use bincode for de/serializing proof-of-sql objects
Recently, proof-of-sql and its dependent services have begun using
bincode with fixed-int encoding and big-endian for
serializing/deserializing over the network. This is because this format
is fairly small and fairly EVM friendly.

This change makes it so ProofPlans are serialized w/ bincode and
VerifiableQueryResults are deserialized w/ bincode accordingly.