fix: upgrade to proof-of-sql 0.75.2
proof-of-sql 0.73.1 was unexpectedly limited in its VarBinary support.
In particular, it did not implement arrow conversions for the VarBinary
column type. These conversions have since been added and released in
proof-of-sql.

This may not be strictly necessary for the SDK, but it is an easy update
to make, and will ensure that the proof-of-sql version is in sync
between the SDK and the prover service.