fix: workaround difference in casing between QueryExpr and storage
It seems that QueryExpr::try_new is probably coercing the table refs and
column refs to lowercase. This is an issue for using the query
commitments provided by substrate as a SchemaAccessor, as all of
substrate's table identifiers and column identifiers are coerced to
uppercase. This PR attempts to work around the issue by implementing a
SchemaAccessor that coerces to uppercase before querying an inner
SchemaAccessor.