feat: parse table refs out of query text
Currently there's a bit of redundancy in client input. The
query_and_verify function accepts the table ref and query text in
separate parameters. However, the table ref should always just be
contained in the query text. This change makes it so the table-ref is
parsed out of the query text, and technically is forward-compatible with
multi-table queries.