feat: implement more Accessors for UppercaseAccessor
It turns out, after we get a successful response from the prover, we
still need the QueryCommitments to act as a MetadataAccessor and
CommitmentAccessor. However, the issue still remains that the
query/proof plan will not necessarily match the all-uppercase chain data
that the query commitments was built from. So, this change makes
UppercaseAccessor implement more proof-of-sql accessor traits. It also
just treats it with a higher level concern in general, giving it its own
module with documentation and tests.